+++ changelog.rst
@@ -18,7 +18,8 @@ Changes by date
 * (enhancement) `direction` parameter is `None` default, to avoid hard-to-track bugs,
 * (debug) debugged `interpolate_raster()` function,
 * (enhancement) kriging data selection - a small refactoring,
-* (docs) Updated `distance` module docs.
+* (docs) Updated `distance` module docs,
+* (enhancement) point kriging functions refactoring and update, better management of a singular matrices and duplicated points,
 
 
 2022-10-22